#164179 NPE

+++ cnd.makeproject/src/org/netbeans/modules/cnd/makeproject/ui/MakeLogicalViewProvider.java
@@ -40,6 +40,7 @@
  */
 package org.netbeans.modules.cnd.makeproject.ui;
 
+import java.awt.EventQueue;
 import java.awt.Toolkit;
 import java.awt.Image;
 import java.awt.color.ColorSpace;
@@ -67,6 +68,7 @@ import javax.swing.SwingUtilities;
 import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
 import org.netbeans.api.project.Project;
+import org.netbeans.api.project.ProjectInformation;
 import org.netbeans.api.project.ProjectUtils;
 import org.netbeans.modules.cnd.api.compilers.Tool;
 import org.netbeans.modules.cnd.api.project.NativeProject;
@@ -77,8 +79,10 @@ import org.netbeans.modules.cnd.makeproject.api.actions.AddExistingItemAction;
 import org.netbeans.modules.cnd.makeproject.api.actions.NewFolderAction;
 import org.netbeans.modules.cnd.makeproject.api.configurations.BooleanConfiguration;
 import org.netbeans.modules.cnd.makeproject.api.configurations.Configuration;
+import org.netbeans.modules.cnd.makeproject.api.configurations.ConfigurationDescriptor.State;
 import org.netbeans.modules.cnd.makeproject.api.configurations.ConfigurationDescriptorProvider;
 import org.netbeans.modules.cnd.makeproject.api.configurations.Configurations;
+import org.netbeans.modules.cnd.makeproject.api.configurations.DevelopmentHostConfiguration;
 import org.netbeans.modules.cnd.makeproject.api.configurations.Folder;
 import org.netbeans.modules.cnd.makeproject.api.configurations.Item;
 import org.netbeans.modules.cnd.makeproject.api.configurations.ItemConfiguration;
@@ -153,14 +157,15 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
     }
 
     public Node createLogicalView() {
+        MakeConfigurationDescriptor configurationDescriptor = getMakeConfigurationDescriptor();
         if (ASYNC_ROOT_NODE) {
             log.fine("creating async root node in EDT? " + SwingUtilities.isEventDispatchThread());// NOI18N
-            return new MakeLogicalViewRootNode(getMakeConfigurationDescriptor().getLogicalFolders());
+            return new MakeLogicalViewRootNode(configurationDescriptor.getLogicalFolders());
         } else {
-            if (getMakeConfigurationDescriptor() == null) {
+            if (configurationDescriptor == null || configurationDescriptor.getState() == State.BROKEN || configurationDescriptor.getConfs().size() == 0) {
                 return new MakeLogicalViewRootNodeBroken();
             } else {
-                return new MakeLogicalViewRootNode(getMakeConfigurationDescriptor().getLogicalFolders());
+                return new MakeLogicalViewRootNode(configurationDescriptor.getLogicalFolders());
             }
         }
     }
@@ -328,6 +333,22 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
         }
     }
 
+    public static void checkForChangedName(final Project project) {
+        SwingUtilities.invokeLater(new Runnable() {
+
+            public void run() {
+                Node rootNode = ProjectTabBridge.getInstance().getExplorerManager().getRootContext();
+                Node root = findProjectNode(rootNode, project);
+                if (root != null) {
+                    ProjectInformation pi = ProjectUtils.getInformation(project);
+                    if (pi != null) { // node will check whether it equals...
+                        root.setDisplayName(pi.getDisplayName());
+                    }
+                }
+            }
+        });
+    }
+
     public static void checkForChangedItems(final Project project, final Folder folder, final Item item) {
         SwingUtilities.invokeLater(new Runnable() {
 
@@ -433,6 +454,23 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
         return new LoadingNode();
     }
 
+    private String getShortDescription() {
+        String prjDirDispName = FileUtil.getFileDisplayName(project.getProjectDirectory());
+        MakeConfigurationDescriptor mkd = getMakeConfigurationDescriptor();
+        if (mkd != null) {
+            MakeConfiguration conf = (MakeConfiguration) mkd.getConfs().getActive();
+            DevelopmentHostConfiguration devHost = (conf == null) ? null : conf.getDevelopmentHost();
+            if (devHost == null || devHost.isLocalhost()) {
+                return NbBundle.getMessage(MakeLogicalViewProvider.class,
+                        "HINT_project_root_node", prjDirDispName); // NOI18N
+            } else {
+                return NbBundle.getMessage(MakeLogicalViewProvider.class,
+                        "HINT_project_root_node_on_host", prjDirDispName, devHost.getDisplayName(true)); // NOI18N
+            }
+        }
+        return prjDirDispName;
+    }
+
     private static class LoadingNode extends AbstractNode {
 
         public LoadingNode() {
@@ -528,19 +566,28 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
 
         @Override
         public String getShortDescription() {
-            String prjDirDispName = FileUtil.getFileDisplayName(project.getProjectDirectory());
-            return NbBundle.getMessage(MakeLogicalViewProvider.class, "HINT_project_root_node", prjDirDispName);
+            return MakeLogicalViewProvider.this.getShortDescription();
         }
+
+        class VisualUpdater implements Runnable {
+            public void run() {
+                fireIconChange();
+                fireOpenedIconChange();
+                fireDisplayNameChange(null, null);
+            }
+        }
+
         /*
          * Something in the folder has changed
          **/
-
         public void stateChanged(ChangeEvent e) {
             brokenLinks = hasBrokenLinks();
             brokenIncludes = hasBrokenIncludes(project);
             updateAnnotationFiles();
-            fireIconChange();
-            fireOpenedIconChange();
+            EventQueue.invokeLater(new VisualUpdater()); // IZ 151257
+//            fireIconChange(); // MakeLogicalViewRootNode
+//            fireOpenedIconChange();
+//            fireDisplayNameChange(null, null);
         }
 
         @Override
@@ -582,18 +629,20 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
 
         @Override
         public Action[] getActions(boolean context) {
+            MakeConfigurationDescriptor descriptor = getMakeConfigurationDescriptor();
+
             // TODO: not clear if we need to call the following method at all
             // but we need to remove remembering the output to prevent memory leak;
             // I think it could be removed
-            if (gotMakeConfigurationDescriptor()) {
-                getMakeConfigurationDescriptor().getLogicalFolders();
+            if (descriptor != null) {
+                descriptor.getLogicalFolders();
             }
 
             List<Action> actions = new ArrayList<Action>();
             // Add standard actions
             Action[] standardActions;
-            MakeConfiguration active = (MakeConfiguration) getMakeConfigurationDescriptor().getConfs().getActive();
-            if (active.isMakefileConfiguration()) { // FIXUP: need better check
+            MakeConfiguration active = (descriptor == null) ? null : (MakeConfiguration) descriptor.getConfs().getActive();
+            if (descriptor == null || active == null || active.isMakefileConfiguration()) { // FIXUP: need better check
                 standardActions = getAdditionalDiskFolderActions();
             } else {
                 standardActions = getAdditionalLogicalFolderActions();
@@ -663,7 +712,7 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
                         ProjectSensitiveActions.projectCommandAction(ActionProvider.COMMAND_BUILD, bundle.getString("LBL_BuildAction_Name"), null), // NOI18N
                         ProjectSensitiveActions.projectCommandAction(ActionProvider.COMMAND_REBUILD, bundle.getString("LBL_RebuildAction_Name"), null), // NOI18N
                         ProjectSensitiveActions.projectCommandAction(ActionProvider.COMMAND_CLEAN, bundle.getString("LBL_CleanAction_Name"), null), // NOI18N
-                        ProjectSensitiveActions.projectCommandAction(MakeActionProvider.COMMAND_BATCH_BUILD, bundle.getString("LBL_BatchBuildAction_Name"), null), // NOI18N
+                        //                        ProjectSensitiveActions.projectCommandAction(MakeActionProvider.COMMAND_BATCH_BUILD, bundle.getString("LBL_BatchBuildAction_Name"), null), // NOI18N
                         ProjectSensitiveActions.projectCommandAction(MakeActionProvider.COMMAND_BUILD_PACKAGE, bundle.getString("LBL_BuildPackagesAction_Name"), null), // NOI18N
                         new RemoteDevelopmentAction(project),
                         new SetConfigurationAction(project),
@@ -696,7 +745,7 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
                         ProjectSensitiveActions.projectCommandAction(ActionProvider.COMMAND_BUILD, bundle.getString("LBL_BuildAction_Name"), null), // NOI18N
                         ProjectSensitiveActions.projectCommandAction(ActionProvider.COMMAND_REBUILD, bundle.getString("LBL_RebuildAction_Name"), null), // NOI18N
                         ProjectSensitiveActions.projectCommandAction(ActionProvider.COMMAND_CLEAN, bundle.getString("LBL_CleanAction_Name"), null), // NOI18N
-                        ProjectSensitiveActions.projectCommandAction(MakeActionProvider.COMMAND_BATCH_BUILD, bundle.getString("LBL_BatchBuildAction_Name"), null), // NOI18N
+                        //                        ProjectSensitiveActions.projectCommandAction(MakeActionProvider.COMMAND_BATCH_BUILD, bundle.getString("LBL_BatchBuildAction_Name"), null), // NOI18N
                         ProjectSensitiveActions.projectCommandAction(MakeActionProvider.COMMAND_BUILD_PACKAGE, bundle.getString("LBL_BuildPackagesAction_Name"), null), // NOI18N
                         new RemoteDevelopmentAction(project),
                         new SetConfigurationAction(project),
@@ -769,8 +818,7 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
 
         @Override
         public String getShortDescription() {
-            String prjDirDispName = FileUtil.getFileDisplayName(project.getProjectDirectory());
-            return NbBundle.getMessage(MakeLogicalViewProvider.class, "HINT_project_root_node", prjDirDispName);
+            return MakeLogicalViewProvider.this.getShortDescription();
         }
     }
 
@@ -894,13 +942,20 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
         }
 
 
+        class VisualUpdater implements Runnable {
+            public void run() {
+                fireIconChange();
+                fireOpenedIconChange();
+            }
+        }
         /*
          * Something in the folder has changed
          **/
         public void stateChanged(ChangeEvent e) {
             updateAnnotationFiles();
-            fireIconChange();
-            fireOpenedIconChange();
+            EventQueue.invokeLater(new VisualUpdater()); // IZ 151257
+//            fireIconChange(); // LogicalFolderNode
+//            fireOpenedIconChange();
         }
 
         public Folder getFolder() {
@@ -1139,9 +1194,10 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
                             copyItemConfigurations(movedItem.getItemConfigurations(), oldConfigurations);
                         }
                     } else {
-                        viewItemNode.getFolder().removeItem(item);
-                        toFolder.addItem(item);
-                        copyItemConfigurations(item.getItemConfigurations(), oldConfigurations);
+                        if (viewItemNode.getFolder().removeItem(item)) {
+                            toFolder.addItem(item);
+                            copyItemConfigurations(item.getItemConfigurations(), oldConfigurations);
+                        }
                     }
                 } else {
                     if (toFolder.isDiskFolder()) {
@@ -1151,8 +1207,9 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
                         String newName = IpeUtils.createUniqueFileName(toFolderPath, itemFO.getName(), itemFO.getExt());
                         FileObject movedFileFO = FileUtil.moveFile(itemFO, toFolderFO, newName);
                     } else if (IpeUtils.isPathAbsolute(item.getPath())) {
-                        viewItemNode.getFolder().removeItem(item);
-                        toFolder.addItem(item);
+                        if (viewItemNode.getFolder().removeItem(item)) {
+                            toFolder.addItem(item);
+                        }
                     } else if (item.getPath().startsWith("..")) { // NOI18N
                         String originalFilePath = FileUtil.toFile(viewItemNode.getFolder().getProject().getProjectDirectory()).getPath();
                         String newFilePath = FileUtil.toFile(toFolder.getProject().getProjectDirectory()).getPath();
@@ -1160,19 +1217,21 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
                         fromNewToOriginal = FilePathAdaptor.normalize(fromNewToOriginal);
                         String newPath = fromNewToOriginal + item.getPath();
                         newPath = IpeUtils.trimDotDot(newPath);
-                        viewItemNode.getFolder().removeItemAction(item);
-                        toFolder.addItemAction(new Item(FilePathAdaptor.normalize(newPath)));
+                        if (viewItemNode.getFolder().removeItemAction(item)) {
+                            toFolder.addItemAction(new Item(FilePathAdaptor.normalize(newPath)));
+                        }
                     } else {
                         Project toProject = toFolder.getProject();
                         FileObject fo = item.getFileObject();
                         FileObject copy = fo.copy(toProject.getProjectDirectory(), fo.getName(), fo.getExt());
                         String newPath = IpeUtils.toRelativePath(FileUtil.toFile(toProject.getProjectDirectory()).getPath(), FileUtil.toFile(copy).getPath());
-                        viewItemNode.getFolder().removeItemAction(item);
-                        fo.delete();
-                        toFolder.addItemAction(new Item(FilePathAdaptor.normalize(newPath)));
+                        if (viewItemNode.getFolder().removeItemAction(item)) {
+                            fo.delete();
+                            toFolder.addItemAction(new Item(FilePathAdaptor.normalize(newPath)));
+                        }
                     }
                 }
-            } else if (type == DnDConstants.ACTION_COPY) {
+            } else if (type == DnDConstants.ACTION_COPY || type == DnDConstants.ACTION_NONE) {
                 // Copy&Paste
                 if (toFolder.getProject() == viewItemNode.getFolder().getProject()) {
                     if ((IpeUtils.isPathAbsolute(item.getPath()) || item.getPath().startsWith("..")) && !toFolder.isDiskFolder()) { // NOI18N
@@ -1600,14 +1659,23 @@ public class MakeLogicalViewProvider implements LogicalViewProvider {
             return excl.getValue();
         }
 
+        class VisualUpdater implements Runnable {
+            public void run() {
+                fireIconChange();
+                fireOpenedIconChange();
+                String displayName = getDisplayName();
+                fireDisplayNameChange(displayName, "");
+                fireDisplayNameChange("", displayName);
+            }
+        }
+
         public void stateChanged(ChangeEvent e) {
-            String displayName = getDisplayName();
-            // do not work:
-            //fireDisplayNameChange(displayName,displayName);
-            fireDisplayNameChange(displayName, "");
-            fireDisplayNameChange("", displayName);
-            fireIconChange();
-            fireOpenedIconChange();
+//            String displayName = getDisplayName();
+//            fireDisplayNameChange(displayName, "");
+//            fireDisplayNameChange("", displayName);
+            EventQueue.invokeLater(new VisualUpdater()); // IZ 151257
+//            fireIconChange(); // ViewItemNode
+//            fireOpenedIconChange();
         }
     }
